requirements.txt
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.2.0
PyMuPDF>=1.24.0
openai>=1.50.0
tiktoken>=0.7.0
gradio>=5.0.0
python-dotenv>=1.0.0
chromadb>=0.5.0
tabulate>=0.9.0
matplotlib>=3.8.0
plotly>=5.18.0
bcrypt>=4.1.0
